+++ src/main/java/dev/xkmc/l2transport/compat/energy/NodalEnergyHandler.java
@@ -8,7 +8,7 @@ public record NodalEnergyHandler(NodalGenericHandler node) implements IEnergySto
 
 	@Override
 	public int receiveEnergy(int maxReceive, boolean simulate) {
-		return TransportHandler.insert(node, EnergyHolder.ENERGY.empty().getCopy(maxReceive), simulate);
+		return (int) TransportHandler.insert(node, EnergyHolder.ENERGY.empty().getCopy(maxReceive), simulate);
 	}
 
 	@Override
