+++ src/main/java/dev/xkmc/l2transport/content/flow/NetworkType.java
@@ -2,10 +2,10 @@ package dev.xkmc.l2transport.content.flow;
 
 public interface NetworkType {
 
-	boolean testConsumption(int avail, int c);
+	boolean testConsumption(long avail, long c);
 
-	boolean shouldContinue(int available, int consumed, int size);
+	boolean shouldContinue(long available, long consumed, long size);
 
-	int provide(int available, int consumed, int size);
+	long provide(long available, long consumed, long size);
 
 }
